allow manual test: servelet call writeXMLs
