+++ src/test/scala/nl.knaw.dans.easy.deposit/SubmitterSpec.scala
@@ -15,6 +15,40 @@
  */
 package nl.knaw.dans.easy.deposit
 
-class SubmitterSpec extends TestSupportFixture {
+import nl.knaw.dans.easy.deposit.docs.DatasetMetadata
+import org.joda.time.DateTime
+import nl.knaw.dans.lib.error._
 
+import scala.util.{ Failure, Success }
+
+class SubmitterSpec extends TestSupportFixture {
+  override def beforeEach(): Unit = {
+    super.beforeEach()
+    clearTestDir()
+    draftsDir.createDirectories()
+  }
+
+  private val draftsDir = testDir / "drafts"
+
+  "submit" should "write 4 files" in {
+    val prologue = """<?xml version='1.0' encoding='UTF-8'?>"""
+    val message = "Lorum ipsum"
+    val datasetMetadata = DatasetMetadata(getManualTestResource("datasetmetadata-from-ui-all.json"))
+      .getOrRecover(e => fail(e))
+      .copy(messageForDataManager = Some(message))
+    val depositDir = DepositDir.create(draftsDir, "user").getOrRecover(e => fail(e))
+    val mdDir = depositDir.getDataFiles.getOrRecover(e => fail(e))
+      .filesMetaData.parent.createIfNotExists(asDirectory = true, createParents = true)
+    depositDir.writeDatasetMetadataJson(datasetMetadata)
+    val oldSize = (mdDir / "dataset.json").size
+    (mdDir.parent / "data" / "text.txt").touch()
+
+    new Submitter(null,null).submit(depositDir) shouldBe a[Failure[_]] // implementation incomplete
+
+    (mdDir / "dataset.json").size shouldBe oldSize // the dataset.json file is not changed
+    (mdDir / "message-from-depositor.txt").contentAsString shouldBe message
+    (mdDir / "agreements.xml").lineIterator.next() shouldBe prologue
+    (mdDir / "dataset.xml").lineIterator.next() shouldBe prologue
+    (mdDir / "files.xml").contentAsString should include("""filepath="data/text.txt""")
+  }
 }
